# test: refactor header inclusion to speed up building