… test/common/upstream/BUILD
… test/common/upstream/eds_speed_test.cc
… test/common/upstream/eds_test.cc
… test/common/upstream/hds_test.cc
… test/common/upstream/logical_dns_cluster_test.cc
… test/common/upstream/original_dst_cluster_test.cc
… test/common/upstream/upstream_impl_test.cc