Signed-off-by: Muge Chen <mugechen@google.com>

Refactor mock class include directive after dividing test/server/mocks.h (#11649)

Building time comparison for affected tests:

|test_name|before|after|
|---|---|---|
|//test/common/upstream:cluster_factory_impl_test|97.051s|80.938s|
|//test/common/upstream:eds_speed_test|102.03s|75.825s|
|//test/common/upstream:eds_test|101.319s|77.988s|
|//test/common/upstream:hds_test|105.666s|74.24s|
|//test/common/upstream:logical_dns_cluster_test|97.184s|78.586s|
|//test/common/upstream:original_dst_cluster_test|116.306s|74.845s|
|//test/common/upstream:transport_socket_matcher_test|99.237s|67.548s|
|//test/common/upstream:upstream_impl_test|101.438s|82.394s|




Commit Message: refactor header inclusion to speed up building
Additional Description:
Risk Level: low
Testing: exsiting tests
Docs Changes: N/A
Release Notes: no
Related Issues: #10917

\cc @ahedberg